Adicionando edição e paginação

+++ src/pages/Books/index.js
@@ -10,30 +10,33 @@ import logoImage from './../../assets/logo.svg';
 export default function Books() {
 
   const [books, setBooks] = useState([]);
+  const [page, setPage] = useState(1);
 
   const userName = localStorage.getItem('userName');
+
   const accessToken = localStorage.getItem('accessToken');
+  const authorization = {
+    headers: {
+      Authorization: `Bearer ${accessToken}`
+    }
+  };
 
   const history = useHistory();
 
   useEffect(() => {
-    api.get('api/v1/Book/asc/20/1', {
-      headers: {
-        Authorization: `Bearer ${accessToken}`
-      }
-    }).then(response => {
-      setBooks(response.data.list)
-    })
+    fetchMoreBooks();
   }, [accessToken]);
+
+  async function fetchMoreBooks() {
+    const response = await api.get(`api/v1/Book/asc/4/${page}`, authorization);
+    setBooks([ ...books, ...response.data.list]);
+    setPage(page + 1);
+  }
 
   async function deleteBook(id, title) {
     try {
       alert(`Tem certeza que deseja deletar o livro ${title}?`)
-      await api.delete(`api/v1/Book/${id}`, {
-        headers: {
-          Authorization: `Bearer ${accessToken}`
-        }
-      });
+      await api.delete(`api/v1/Book/${id}`, authorization);
 
       setBooks(books.filter(book => book.id !== id));
     } catch (err) {
@@ -41,13 +44,17 @@ export default function Books() {
     }
   }
 
+  async function editBook(id) {
+    try {
+      history.push(`book/new/${id}`);
+    } catch (err) {
+      alert('Edit book failed! Try again.')
+    }
+  }
+
   async function logout() {
     try {
-      await api.get('api/v1/Auth/revoke', {
-        headers: {
-          Authorization: `Bearer ${accessToken}`
-        }
-      });
+      await api.get('api/v1/Auth/revoke', authorization);
 
       localStorage.clear();
       history.push('/');
@@ -61,7 +68,7 @@ export default function Books() {
       <header>
         <img src={logoImage} alt="Erudio" />
         <span>Welcome, <strong>{userName.toLowerCase()}</strong>!</span>
-        <Link className="button" to="book/new">Add New Book</Link>
+        <Link className="button" to="book/new/0">Add New Book</Link>
         <button type="button" onClick={logout}>
           <FiPower size={18} color="#251fc5" />
         </button>
@@ -80,7 +87,7 @@ export default function Books() {
             <strong>Release Date:</strong>
             <p>{Intl.DateTimeFormat('pt-BR').format(new Date(book.launchDate))}</p>
 
-            <button type="button">
+            <button type="button" onClick={() => editBook(book.id)}>
               <FiEdit size={20} color="#251fc5" />
             </button>
             <button type="button" onClick={() => deleteBook(book.id, book.title)}>
@@ -89,6 +96,7 @@ export default function Books() {
           </li>
         ))}
       </ul>
+      <button className="button" onClick={fetchMoreBooks} type="button">Load More</button>
     </div>
   )
 }
